refactor BrowserError to be a named export
