+++ src/utils/page-functions.js
@@ -33,6 +33,7 @@ export async function createPageFunctions(page, sourceFiles, babelPresets) {
     }
 
     if (pageFunctions[fnName]) {
+      // eslint-disable-next-line no-console
       console.warn(`A page function with name '${fnName}' already exists, the old one will be overwritten`)
     }
 
@@ -178,7 +179,7 @@ export async function getCacheEntry(filePath = '') {
 }
 
 export async function checkNodeModulesPath(modulesPath) {
-  let pathExists = await exists(modulesPath)
+  const pathExists = await exists(modulesPath)
 
   if (pathExists) {
     nodeModulesPath = modulesPath
